# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Aapater API",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
        "/api/v1/pet": {
            "get": {
                "tags": [
                    "pet"
                ],
                "summary": "Get Pet by animal_kind",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search animal_kind:{貓/狗}",
                        "name": "animal_kind",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Pet"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "pet"
                ],
                "summary": "Add Pet",
                "parameters": [
                    {
                        "description": "Pet",
                        "name": "Pet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Pet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/apis.HttpResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/version": {
            "get": {
                "tags": [
                    "version"
                ],
                "summary": "Get version",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apis.HttpResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.Pet": {
            "type": "object",
            "properties": {
                "animal_age": {
                    "type": "string"
                },
                "animal_id": {
                    "type": "string"
                },
                "animal_kind": {
                    "type": "string"
                },
                "animal_sex": {
                    "type": "string"
                }
            }
        }
    }
}